add pytorch version
