+++ tensorflow/imagenet_utils.py
@@ -1,0 +1,342 @@
+# 2020.02.26-Changed for utils for testing GhostNet on ImageNet
+#            Huawei Technologies Co., Ltd. <foss@huawei.com>
+# modified from https://github.com/tensorpack/tensorpack/blob/master/examples/ImageNetModels/imagenet_utils.py
+import cv2
+import numpy as np
+import multiprocessing
+import tensorflow as tf
+from tensorflow.python.framework import ops
+from abc import abstractmethod
+
+from tensorpack import imgaug, dataset, ModelDesc
+from tensorpack.dataflow import (
+    AugmentImageComponent, PrefetchDataZMQ, MapData,
+    BatchData, MultiThreadMapData)
+from tensorpack.predict import PredictConfig, SimpleDatasetPredictor
+from tensorpack.utils.stats import RatioCounter
+from tensorpack.models import regularize_cost
+from tensorpack.tfutils.summary import add_moving_summary
+from tensorpack.utils import logger
+
+
+class GoogleNetResize(imgaug.ImageAugmentor):
+    """
+    crop 8%~100% of the original image
+    See `Going Deeper with Convolutions` by Google.
+    """
+    def __init__(self, crop_area_fraction=0.08,
+                 aspect_ratio_low=0.75, aspect_ratio_high=1.333,
+                 target_shape=224):
+        self._init(locals())
+
+    def _augment(self, img, _):
+        h, w = img.shape[:2]
+        area = h * w
+        for _ in range(10):
+            targetArea = self.rng.uniform(self.crop_area_fraction, 1.0) * area
+            aspectR = self.rng.uniform(self.aspect_ratio_low, self.aspect_ratio_high)
+            ww = int(np.sqrt(targetArea * aspectR) + 0.5)
+            hh = int(np.sqrt(targetArea / aspectR) + 0.5)
+            if self.rng.uniform() < 0.5:
+                ww, hh = hh, ww
+            if hh <= h and ww <= w:
+                x1 = 0 if w == ww else self.rng.randint(0, w - ww)
+                y1 = 0 if h == hh else self.rng.randint(0, h - hh)
+                out = img[y1:y1 + hh, x1:x1 + ww]
+                out = cv2.resize(out, (self.target_shape, self.target_shape), interpolation=cv2.INTER_CUBIC)
+                return out
+        out = imgaug.ResizeShortestEdge(self.target_shape, interp=cv2.INTER_CUBIC).augment(img)
+        out = imgaug.CenterCrop(self.target_shape).augment(out)
+        return out
+
+def fbresnet_augmentor(isTrain):
+    """
+    Augmentor used in fb.resnet.torch, for BGR images in range [0,255].
+    """
+    if isTrain:
+        augmentors = [
+            GoogleNetResize(),
+            # It's OK to remove the following augs if your CPU is not fast enough.
+            # Removing brightness/contrast/saturation does not have a significant effect on accuracy.
+            # Removing lighting leads to a tiny drop in accuracy.
+            imgaug.RandomOrderAug(
+                [imgaug.BrightnessScale((0.6, 1.4), clip=False),
+                 imgaug.Contrast((0.6, 1.4), clip=False),
+                 imgaug.Saturation(0.4, rgb=False),
+                 # rgb-bgr conversion for the constants copied from fb.resnet.torch
+                 imgaug.Lighting(0.1,
+                                 eigval=np.asarray(
+                                     [0.2175, 0.0188, 0.0045][::-1]) * 255.0,
+                                 eigvec=np.array(
+                                     [[-0.5675, 0.7192, 0.4009],
+                                      [-0.5808, -0.0045, -0.8140],
+                                      [-0.5836, -0.6948, 0.4203]],
+                                     dtype='float32')[::-1, ::-1]
+                                 )]),
+            imgaug.Flip(horiz=True),
+        ]
+    else:
+        augmentors = [
+            imgaug.ResizeShortestEdge(256, cv2.INTER_CUBIC),
+            imgaug.CenterCrop((224, 224)),
+        ]
+    return augmentors
+
+
+def get_imagenet_dataflow(
+        datadir, name, batch_size,
+        augmentors, meta_dir=None, parallel=None):
+    """
+    See explanations in the tutorial:
+    http://tensorpack.readthedocs.io/en/latest/tutorial/efficient-dataflow.html
+    """
+    assert name in ['train', 'val', 'test']
+    assert datadir is not None
+    assert isinstance(augmentors, list)
+    isTrain = name == 'train'
+    
+    #parallel = 1
+    
+    if parallel is None:
+        parallel = min(40, multiprocessing.cpu_count() // 2)  # assuming hyperthreading
+    if isTrain:
+        ds = dataset.ILSVRC12(datadir, name, meta_dir=meta_dir, shuffle=True)
+        ds = AugmentImageComponent(ds, augmentors, copy=False)
+        if parallel < 16:
+            logger.warn("DataFlow may become the bottleneck when too few processes are used.")
+        ds = PrefetchDataZMQ(ds, parallel)
+        ds = BatchData(ds, batch_size, remainder=False)
+    else:
+        ds = dataset.ILSVRC12Files(datadir, name, meta_dir= meta_dir, shuffle=False)
+        aug = imgaug.AugmentorList(augmentors)
+
+        def mapf(dp):
+            fname, cls = dp
+            im = cv2.imread(fname, cv2.IMREAD_COLOR)
+            im = aug.augment(im)
+            return im, cls
+        ds = MultiThreadMapData(ds, parallel, mapf, buffer_size=2000, strict=True)
+        ds = BatchData(ds, batch_size, remainder=True)
+        ds = PrefetchDataZMQ(ds, 1)
+    return ds
+
+
+def eval_on_ILSVRC12(model, sessinit, dataflow):
+    pred_config = PredictConfig(
+        model=model,
+        session_init=sessinit,
+        input_names=['input', 'label'],
+        output_names=['wrong-top1', 'wrong-top5']
+    )
+    pred = SimpleDatasetPredictor(pred_config, dataflow)
+    acc1, acc5 = RatioCounter(), RatioCounter()
+    for top1, top5 in pred.get_result():
+        batch_size = top1.shape[0]
+        acc1.feed(top1.sum(), batch_size)
+        acc5.feed(top5.sum(), batch_size)
+    print("Top1 Error: {}".format(acc1.ratio))
+    print("Top5 Error: {}".format(acc5.ratio))
+
+
+class ImageNetModel(ModelDesc):
+    image_shape = 224
+    lr = 0.1
+
+    """
+    uint8 instead of float32 is used as input type to reduce copy overhead.
+    It might hurt the performance a liiiitle bit.
+    The pretrained models were trained with float32.
+    """
+    image_dtype = tf.float32
+
+    """
+    Either 'NCHW' or 'NHWC'
+    """
+    data_format = 'NCHW'
+
+    """
+    Whether the image is BGR or RGB. If using DataFlow, then it should be BGR.
+    """
+    image_bgr = True
+
+    weight_decay = 4e-5
+    label_smoothing = 0.0
+
+    """
+    To apply on normalization parameters, use '.*/gamma|.*/beta'
+    to apply on depthwise, add '.*/DW'
+    """
+    weight_decay_pattern = '.*/W|.*/M|.*/WB|.*/weights'
+
+    """
+    Scale the loss, for whatever reasons (e.g., gradient averaging, fp16 training, etc)
+    """
+    loss_scale = 1.
+
+    def inputs(self):        
+        labels = tf.placeholder(tf.int32, [None], 'label')
+        return [tf.placeholder(self.image_dtype, [None, self.image_shape, self.image_shape, 3], 'input'),
+               labels]
+
+    def build_graph(self, image, label):
+        image = ImageNetModel.image_preprocess(image, bgr=self.image_bgr)
+        assert self.data_format in ['NCHW', 'NHWC']
+        if self.data_format == 'NCHW':
+            image = tf.transpose(image, [0, 3, 1, 2])
+
+        logits = self.get_logits(image)
+        print('self.label_smoothing', self.label_smoothing)
+        loss = ImageNetModel.compute_loss_and_error(logits, label, self.label_smoothing)
+
+        if self.weight_decay > 0:
+            wd_loss = regularize_cost(self.weight_decay_pattern,
+                                      tf.contrib.layers.l2_regularizer(self.weight_decay),
+                                      name='l2_regularize_loss')
+            add_moving_summary(loss, wd_loss)
+            total_cost = tf.add_n([loss, wd_loss], name='cost')
+        else:
+            total_cost = tf.identity(loss, name='cost')
+            add_moving_summary(total_cost)
+
+        if self.loss_scale != 1.:
+            logger.info("Scaling the total loss by {} ...".format(self.loss_scale))
+            return total_cost * self.loss_scale
+        else:
+            return total_cost
+
+    @abstractmethod
+    def get_logits(self, image):
+        """
+        Args:
+            image: 4D tensor of ``self.input_shape`` in ``self.data_format``
+
+        Returns:
+            Nx#class logits
+        """
+
+    def optimizer(self):
+        lr = tf.get_variable('learning_rate', initializer=self.lr, trainable=False)
+        tf.summary.scalar('learning_rate-summary', lr)
+        return tf.train.MomentumOptimizer(lr, 0.9, use_nesterov=True)
+        #return tf.train.RMSPropOptimizer(lr, momentum=0.9) #tf.train.MomentumOptimizer(lr, 0.9, use_nesterov=True)
+
+    @staticmethod
+    def image_preprocess(image, bgr=True):
+        with tf.name_scope('image_preprocess'):
+            if image.dtype.base_dtype != tf.float32:
+                image = tf.cast(image, tf.float32)
+            image = image * (1.0 / 255)
+
+            mean = [0.485, 0.456, 0.406]    # rgb
+            std = [0.229, 0.224, 0.225]
+            if bgr:
+                mean = mean[::-1]
+                std = std[::-1]
+            image_mean = tf.constant(mean, dtype=tf.float32)
+            image_std = tf.constant(std, dtype=tf.float32)
+            image = (image - image_mean) / image_std
+            return image
+
+    @staticmethod
+    def compute_loss_and_error(logits, label, label_smoothing):
+        loss = sparse_softmax_cross_entropy(
+                logits=logits, labels=label,
+                label_smoothing = label_smoothing,
+                weights=1.0)
+        loss = tf.reduce_mean(loss, name='xentropy-loss')
+
+        def prediction_incorrect(logits, label, topk=1, name='incorrect_vector'):
+            with tf.name_scope('prediction_incorrect'):
+                x = tf.logical_not(tf.nn.in_top_k(logits, label, topk))
+            return tf.cast(x, tf.float32, name=name)
+        
+        if label.shape.ndims > 1:
+            label = tf.cast(tf.argmax(label, axis=1), tf.int32)
+        wrong = prediction_incorrect(logits, label, 1, name='wrong-top1')
+        add_moving_summary(tf.reduce_mean(wrong, name='train-error-top1'))
+
+        wrong = prediction_incorrect(logits, label, 5, name='wrong-top5')
+        add_moving_summary(tf.reduce_mean(wrong, name='train-error-top5'))
+        return loss
+
+def sparse_softmax_cross_entropy(
+        labels,
+        logits,
+        weights=1.0,
+        label_smoothing=0.0,
+        scope=None,
+        loss_collection=ops.GraphKeys.LOSSES,
+        reduction=tf.losses.Reduction.SUM_BY_NONZERO_WEIGHTS):
+    """Cross-entropy loss using `tf.nn.sparse_softmax_cross_entropy_with_logits`.
+    `weights` acts as a coefficient for the loss. If a scalar is provided,
+    then the loss is simply scaled by the given value. If `weights` is a
+    tensor of shape [`batch_size`], then the loss weights apply to each
+    corresponding sample.
+    Args:
+        labels: `Tensor` of shape `[d_0, d_1, ..., d_{r-1}]` (where `r` is rank of
+        `labels` and result) and dtype `int32` or `int64`. Each entry in `labels`
+        must be an index in `[0, num_classes)`. Other values will raise an
+        exception when this op is run on CPU, and return `NaN` for corresponding
+        loss and gradient rows on GPU.
+        logits: Unscaled log probabilities of shape
+        `[d_0, d_1, ..., d_{r-1}, num_classes]` and dtype `float32` or `float64`.
+        weights: Coefficients for the loss. This must be scalar or broadcastable to
+        `labels` (i.e. same rank and each dimension is either 1 or the same).
+        scope: the scope for the operations performed in computing the loss.
+        loss_collection: collection to which the loss will be added.
+        reduction: Type of reduction to apply to loss.
+    Returns:
+        Weighted loss `Tensor` of the same type as `logits`. If `reduction` is
+        `NONE`, this has the same shape as `labels`; otherwise, it is scalar.
+    Raises:
+        ValueError: If the shapes of `logits`, `labels`, and `weights` are
+        incompatible, or if any of them are None.
+    """
+    if labels is None:
+        raise ValueError("labels must not be None.")
+    if logits is None:
+        raise ValueError("logits must not be None.")
+    with tf.name_scope(scope, "sparse_softmax_cross_entropy_loss",
+                       (logits, labels, weights)) as scope:
+
+        if labels.shape.ndims == 1:
+            loss = tf.nn.sparse_softmax_cross_entropy_with_logits(
+                logits=logits, labels=labels, name="xentropy")
+        else:
+            loss = tf.nn.softmax_cross_entropy_with_logits(
+                labels=labels, logits=logits, name="xentropy")            
+
+        loss = tf.losses.compute_weighted_loss(
+            loss, weights, scope, loss_collection, reduction=reduction)
+
+        # Label smoothing.
+        smooth_loss = 0.
+        if label_smoothing > 0:
+            # Label smoothing loss: sum of logits * weight.
+            loss = tf.scalar_mul(1. - label_smoothing, loss)
+            aux_log_softmax = -tf.nn.log_softmax(logits)
+
+            smooth_loss = tf.losses.compute_weighted_loss(
+                aux_log_softmax, label_smoothing * weights,
+                'label_smoothing', loss_collection, reduction=reduction)
+
+        return loss + smooth_loss
+
+
+if __name__ == '__main__':
+    import argparse
+    from tensorpack.dataflow import TestDataSpeed
+    parser = argparse.ArgumentParser()
+    parser.add_argument('--data', required=True)
+    parser.add_argument('--batch', type=int, default=32)
+    parser.add_argument('--aug', choices=['train', 'val'], default='val')
+    args = parser.parse_args()
+
+    if args.aug == 'val':
+        augs = fbresnet_augmentor(False)
+    elif args.aug == 'train':
+        augs = fbresnet_augmentor(True)
+    df = get_imagenet_dataflow(
+        args.data, 'train', args.batch, augs)
+    # For val augmentor, Should get >100 it/s (i.e. 3k im/s) here on a decent E5 server.
+    TestDataSpeed(df).start()
